app/django/requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
Django==6.0
django-filter==25.2
djangorestframework==3.16.1
dsnparser==0.0.3
greenlet==3.3.0
Markdown==3.10
mysql==0.0.3
mysqlclient==2.2.7
pydantic==2.12.5
pydantic_core==2.41.5
SQLAlchemy==2.0.45
sqlmodel==0.0.31
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
